requirements.txt
zeroth-learn
matplotlib
dynamics
pandas
numpy
jax
